tambah edit profil dan permission

+++ resources/views/admin/admin/admins.blade.php
@@ -22,8 +22,10 @@
 			</div>
 		</div>
 		<div class="pull-right">
-			<a href="{{route('admin.create')}}" class="btn btn-primary">Tambah User</a>
-			<button id="bulkDelete" class="btn btn-danger">Hapus</button>
+			@if(Auth::User()->level == 'admin')
+				<a href="{{route('admin.create')}}" class="btn btn-primary">Tambah User</a>
+				<button id="bulkDelete" class="btn btn-danger">Hapus</button>
+			@endif
 		</div>
 	</div>
 	<div class="panel-body" id="data-body">
@@ -38,7 +40,13 @@
 		<table id="datatable" class="table table-hover table-bordered">
 			<thead>
 				<tr>
-					<th class="text-center valign-middle"><input type="checkbox" id="togglecb"></th>
+					<th class="text-center valign-middle">
+						@if(Auth::User()->level == 'admin')
+							<input type="checkbox" id="togglecb">
+						@else
+							#
+						@endif
+					</th>
 					<th width="30%">Nama</th>
 					<th>No. Induk</th>
 					<th>No. Telepon</th>
@@ -49,18 +57,28 @@
 			<tbody>
 				@foreach( $data['users'] as $user )
 					<tr data-id="{{$user->id}}">
-						<td class="text-center valign-middle"><input type="checkbox" class="cbRow" value="{{ $user->id }}" name="id[]"></td>
-						<td data-search="{{ $user->name }}"><a href="{{route('admin.edit', $user->id)}}" title="Edit user">{{ $user->nama }}</a></td>
+						<td class="text-center valign-middle">
+							@if(Auth::User()->level == 'admin')
+								<input type="checkbox" class="cbRow" value="{{ $user->id }}" name="id[]">
+							@else 
+								{{$user->id}}
+							@endif
+						</td>
+						<td data-search="{{ $user->nama }}"><a href="{{route('admin.edit', $user->id)}}" title="Edit user">{{ $user->nama }}</a></td>
 						<td data-search="{{ $user->noinduk }}">{{ $user->noinduk }}</td>
 						<td data-search="{{ $user->telepon_1 }} {{ $user->telepon_2 }}">{{ $user->telepon_1 }} {{ ! empty( $user->telepon_2) ? '/ ' . $user->telepon_2 : '' }}</td>
 						<td data-order="{{ $user->level }}">{{ ucfirst( $user->level) }}</td>
 						<td>
-							<a href="{{route('admin.edit', $user->id) }}" title="Edit user" class="btn btn-primary"><i class="fa fa-pencil"></i></a>
-							<form style="display: inline;" action="{{route('admin.destroy', $user->id)}}" method="POST">
-								<input type="hidden" name="_method" value="DELETE">
-								{{ csrf_field() }}
-								<button type="submit" title="Hapus user" class="btn btn-danger"><i class="fa fa-trash-o"></i></button>
-							</form>
+							@if(Auth::User()->level == 'admin')
+								<a href="{{route('admin.edit', $user->id) }}" title="Edit user" class="btn btn-primary"><i class="fa fa-pencil"></i></a>
+								<form style="display: inline;" action="{{route('admin.destroy', $user->id)}}" method="POST">
+									<input type="hidden" name="_method" value="DELETE">
+									{{ csrf_field() }}
+									<button type="submit" title="Hapus user" class="btn btn-danger"><i class="fa fa-trash-o"></i></button>
+								</form>
+							@else 
+								-
+							@endif
 						</td>
 					</tr>
 				@endforeach
